% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{coef.HCgglasso}
\alias{coef.HCgglasso}
\title{Get coefficients from a \code{\link{HCgglasso}} object}
\usage{
\method{coef}{HCgglasso}(object, s = NULL, ...)
}
\arguments{
\item{object}{\code{\link{HCgglasso}} object}

\item{s}{values of lambda. If NULL, use values from object}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with estimated coefficients for given values of s.
}
\description{
Get coefficients from a \code{\link{HCgglasso}} object
}
\author{
Quentin Grimonprez
}
\seealso{
\link{HCgglasso}, \link{predict.HCgglasso}
}

