% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{listToMatrix}
\alias{listToMatrix}
\title{Obtain a sparse matrix of the coefficients of the path}
\usage{
listToMatrix(x, row = c("covariates", "lambda"))
}
\arguments{
\item{x}{\code{\link{HCgglasso}} object}

\item{row}{"lambda" or "covariates". If row="covariates", each row of the output matrix represents a covariates else ff row="lambda", it represents a value of lambda.}
}
\value{
a sparse matrix containing the estimated coefficients for different values of lambda
}
\description{
Obtain a sparse matrix of the coefficients of the path
}
\details{
This functions can be used with a \code{\link{HCgglasso}} object to obtain a matrix with all estimated coefficients for the p original variables.
In case of overlapping groups, coefficients from repeated variables are summed.
}
\examples{
X=simuBlockGaussian(50,12,5,0.7)
y=drop(X[,c(2,7,12)]\%*\%c(2,2,-2)+rnorm(50,0,0.5))
res=HCgglasso(X,y)
beta=listToMatrix(res)

}
\seealso{
\link{HCgglasso}, \link{overlapgglasso}
}

