% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot.HCgglasso}
\alias{plot.HCgglasso}
\title{Plot the path obtained from \code{\link{HCgglasso}} function}
\usage{
\method{plot}{HCgglasso}(x, log.lambda = FALSE, lambda.lines = FALSE, ...)
}
\arguments{
\item{x}{\code{\link{HCgglasso}} object}

\item{log.lambda}{If TRUE, use log(lambda) instead of lambda in abscissa}

\item{lambda.lines}{if TRUE, add vertical lines at lambda values}

\item{...}{Other parameters for plot function}
}
\description{
Plot the path obtained from \code{\link{HCgglasso}} function
}
\examples{
\dontrun{
X=simuBlockGaussian(50,12,5,0.7)
y=drop(X[,c(2,7,12)]\%*\%c(2,2,-2)+rnorm(50,0,0.5))
res=HCgglasso(X,y)
plot(res) 
}

}
\seealso{
\link{HCgglasso}
}

