% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cv.HCgglasso}
\alias{predict.cv.HCgglasso}
\title{Predict fitted values from a \code{\link{cv.HCgglasso}} object}
\usage{
\method{predict}{cv.HCgglasso}(object, newx = NULL, s = c("lambda.1se",
  "lambda.min"), type = c("fit", "coefficients"), ...)
}
\arguments{
\item{object}{\code{\link{cv.HCgglasso}} object}

\item{newx}{matrix with new individuals for prediction. If type="coefficients", the parameter has to be NULL}

\item{s}{Either "lambda.1se" or "lambda.min"}

\item{type}{if "fit", return the fitted values for each values of s, if "coefficients", return the estimated coefficients for each s}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with fitted values or estimated coefficients for given values of s.
}
\description{
Predict fitted values from a \code{\link{cv.HCgglasso}} object
}
\author{
Quentin Grimonprez
}
\seealso{
\link{cv.HCgglasso}
}

