% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.HCgglasso}
\alias{predict.HCgglasso}
\title{Predict fitted values from a \code{\link{HCgglasso}} object}
\usage{
\method{predict}{HCgglasso}(object, newx = NULL, s = NULL, type = c("fit",
  "coefficients"), ...)
}
\arguments{
\item{object}{\code{\link{HCgglasso}} object}

\item{newx}{matrix with new individuals for prediction. If type="coefficients", the parameter has to be NULL}

\item{s}{values of lambda. If NULL, use values from object}

\item{type}{if "fit", return the fitted values for each values of s, if "coefficients", return the estimated coefficients for each s}

\item{...}{Not used. Other arguments to predict.}
}
\value{
A matrix with fitted values or estimated coefficients for given values of s.
}
\description{
Predict fitted values from a \code{\link{HCgglasso}} object
}
\author{
original code from \pkg{gglasso} package Author: Yi Yang <yiyang@umn.edu>, Hui Zou <hzou@stat.umn.edu>
}
\seealso{
\link{HCgglasso}
}

